requirements.txt
pdf2image
pillow
pathlib
numpy
scikit-learn